+++ src/test/java/kitchenpos/product/domain/ProductTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.product.domain;
 
 import org.junit.jupiter.api.Test;
